Split models module into prf and cf submodules
