+++ src/prfmodel/models/__init__.py
@@ -4,20 +4,20 @@ from .base import BaseComposite
 from .base import BaseEncoder
 from .base import BaseResponse
 from .base import ModelProtocol
-from .css import Gaussian2DCSSPRFModel
-from .css import init_css_from_gaussian
-from .div_norm import DivNormGaussian2DPRFModel
-from .div_norm import DivNormPRFModel
-from .div_norm import init_dn_from_dog
-from .div_norm import init_dn_from_gaussian
-from .dog import DoG2DPRFModel
-from .dog import init_dog_from_gaussian
-from .encoding import CFStimulusEncoder
-from .encoding import CompressiveEncoder
-from .encoding import PRFStimulusEncoder
-from .encoding import encode_prf_response
-from .gaussian import Gaussian2DPRFModel
-from .gaussian import GaussianCFModel
+from .cf.gaussian import GaussianCFModel
+from .cf.stimulus_encoding import CFStimulusEncoder
+from .prf.css import Gaussian2DCSSPRFModel
+from .prf.css import init_css_from_gaussian
+from .prf.div_norm import DivNormGaussian2DPRFModel
+from .prf.div_norm import DivNormPRFModel
+from .prf.div_norm import init_dn_from_dog
+from .prf.div_norm import init_dn_from_gaussian
+from .prf.dog import DoG2DPRFModel
+from .prf.dog import init_dog_from_gaussian
+from .prf.gaussian import Gaussian2DPRFModel
+from .prf.stimulus_encoding import CompressiveEncoder
+from .prf.stimulus_encoding import PRFStimulusEncoder
+from .prf.stimulus_encoding import encode_prf_response
 
 __all__ = [
     "BaseComposite",
