+++ src/prfmodel/models/cf/__init__.py
@@ -1,0 +1,13 @@
+"""Connective field models."""
+
+from .composite import SimpleCFModel
+from .gaussian import GaussianCFModel
+from .gaussian import GaussianCFResponse
+from .stimulus_encoding import CFStimulusEncoder
+
+__all__ = [
+    "CFStimulusEncoder",
+    "GaussianCFModel",
+    "GaussianCFResponse",
+    "SimpleCFModel",
+]
