+++ src/prfmodel/models/prf/composite.py
@@ -7,18 +7,16 @@ from prfmodel._docstring import doc
 from prfmodel.impulse import DerivativeTwoGammaImpulse
 from prfmodel.impulse import convolve_prf_impulse_response
 from prfmodel.impulse.base import BaseImpulse
+from prfmodel.models.base import BaseComposite
+from prfmodel.models.base import BaseEncoder
+from prfmodel.models.base import BaseResponse
 from prfmodel.scaling import BaselineAmplitude
 from prfmodel.scaling import DoGAmplitude
 from prfmodel.scaling.base import BaseScaling
-from prfmodel.stimuli import CFStimulus
 from prfmodel.stimuli import PRFStimulus
 from prfmodel.typing import Tensor
 from prfmodel.utils import get_dtype
-from .base import BaseComposite
-from .base import BaseEncoder
-from .base import BaseResponse
-from .encoding import CFStimulusEncoder
-from .encoding import PRFStimulusEncoder
+from .stimulus_encoding import PRFStimulusEncoder
 
 
 class SimplePRFModel(BaseComposite[PRFStimulus]):
@@ -276,76 +274,3 @@ class CenterSurroundPRFModel(BaseComposite[PRFStimulus]):
 
         # When scaling_model=None, return a simple subtraction (resp1 - resp2)
         return stacked[:, 0] - stacked[:, 1]
-
-
-class SimpleCFModel(BaseComposite[CFStimulus]):
-    """
-    Simple composite connective field model.
-
-    This is a generic class that combines a connective field and temporal response.
-
-    Parameters
-    ----------
-    %(model_cf)s
-    %(model_scaling)s
-
-    Notes
-    -----
-    The simple composite model follows three steps:
-
-    1. The connective field response model makes a prediction for the stimulus distance matrix.
-    2. The connective field response is encoded with the source response.
-    3. The temporal model modifies the encoded response.
-
-    """
-
-    def __init__(
-        self,
-        cf_model: BaseResponse,
-        encoding_model: BaseEncoder | type[BaseEncoder] = CFStimulusEncoder,
-        scaling_model: BaseScaling | type[BaseScaling] | None = BaselineAmplitude,
-    ):
-        if encoding_model is not None and isinstance(encoding_model, type):
-            encoding_model = encoding_model()
-
-        if scaling_model is not None and isinstance(scaling_model, type):
-            scaling_model = scaling_model()
-
-        super().__init__(
-            cf_model=cf_model,
-            encoding_model=encoding_model,
-            scaling_model=scaling_model,
-        )
-
-    @doc
-    def __call__(
-        self,
-        stimulus: CFStimulus,
-        parameters: pd.DataFrame,
-        dtype: str | None = None,
-    ) -> Tensor:
-        """
-        Predict a simple connective field model response to a stimulus.
-
-        Parameters
-        ----------
-        %(stimulus_cf)s
-        %(parameters)s
-        %(dtype)s
-
-        Returns
-        -------
-        %(predicted_response_2d)s
-
-        """
-        dtype = get_dtype(dtype)
-        cf_model = cast("BaseResponse", self.models["cf_model"])
-        response = cf_model(stimulus, parameters, dtype=dtype)
-        encoding_model = cast("BaseEncoder", self.models["encoding_model"])
-        response = encoding_model(stimulus, response, parameters, dtype=dtype)
-
-        if self.models["scaling_model"] is not None:
-            temporal_model = cast("BaseScaling", self.models["scaling_model"])
-            response = temporal_model(response, parameters, dtype=dtype)
-
-        return response
