+++ src/prfmodel/models/prf/div_norm.py
@@ -6,16 +6,16 @@ from keras import ops
 from prfmodel.impulse import DerivativeTwoGammaImpulse
 from prfmodel.impulse import convolve_prf_impulse_response
 from prfmodel.impulse.base import BaseImpulse
+from prfmodel.models.base import BaseComposite
+from prfmodel.models.base import BaseEncoder
+from prfmodel.models.base import BaseResponse
 from prfmodel.scaling import DivNormAmplitude
 from prfmodel.scaling.base import BaseScaling
 from prfmodel.stimuli.prf import PRFStimulus
 from prfmodel.typing import Tensor
 from prfmodel.utils import get_dtype
-from .base import BaseComposite
-from .base import BaseEncoder
-from .base import BaseResponse
-from .encoding import PRFStimulusEncoder
 from .gaussian import Gaussian2DPRFResponse
+from .stimulus_encoding import PRFStimulusEncoder
 
 
 class DivNormPRFModel(BaseComposite[PRFStimulus]):
