+++ src/prfmodel/models/prf/gaussian.py
@@ -1,7 +1,6 @@
 """Gaussian response models."""
 
 import math
-import numpy as np
 import pandas as pd
 from keras import ops
 from prfmodel._docstring import doc
@@ -9,21 +8,18 @@ from prfmodel.exceptions import BatchDimensionError
 from prfmodel.exceptions import ShapeError
 from prfmodel.impulse import DerivativeTwoGammaImpulse
 from prfmodel.impulse.base import BaseImpulse
+from prfmodel.models.base import BaseEncoder
+from prfmodel.models.base import BaseResponse
 from prfmodel.scaling import BaselineAmplitude
 from prfmodel.scaling.base import BaseScaling
-from prfmodel.stimuli import CFStimulus
 from prfmodel.stimuli import GridDimensionsError
 from prfmodel.stimuli import PRFStimulus
 from prfmodel.typing import Tensor
 from prfmodel.utils import _EXPECTED_NDIM
 from prfmodel.utils import convert_parameters_to_tensor
 from prfmodel.utils import get_dtype
-from .base import BaseEncoder
-from .base import BaseResponse
-from .composite import SimpleCFModel
 from .composite import SimplePRFModel
-from .encoding import CFStimulusEncoder
-from .encoding import PRFStimulusEncoder
+from .stimulus_encoding import PRFStimulusEncoder
 
 
 class GridMuDimensionsError(Exception):
@@ -227,84 +223,6 @@ class Gaussian2DPRFResponse(BaseResponse[PRFStimulus]):
         grid = ops.convert_to_tensor(stimulus.grid, dtype=dtype)
 
         return predict_gaussian_response(grid, mu, sigma)
-
-
-class GaussianCFResponse(BaseResponse[CFStimulus]):
-    """
-    Gaussian connective field response model.
-
-    Predicts a response to a stimulus distance matrix.
-    The model has two parameters: `center_index` is the index of the row in the stimulus distance matrix that is the
-    center of the Gaussian; `sigma` for the width of the Gaussian.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> import pandas as pd
-    >>> from prfmodel.stimuli.cf import CFStimulus
-    >>> num_source_units, num_frames = 10, 20
-    >>> distances = np.abs(
-    ...     np.arange(num_source_units, dtype=float)[:, None]
-    ...     - np.arange(num_source_units, dtype=float)[None, :]
-    ... )
-    >>> source_response = np.ones((num_source_units, num_frames))
-    >>> stimulus = CFStimulus(
-    ...     distance_matrix=distances,
-    ...     source_response=source_response
-    ... )
-    >>> # Define parameters for 2 target units
-    >>> params = pd.DataFrame({
-    ...     "center_index": [0, 5],
-    ...     "sigma": [1.0, 2.0]
-    ... })
-    >>> model = GaussianCFResponse()
-    >>> resp = model(stimulus, params)
-    >>> print(resp.shape)  # (num_units, num_source_units)
-    (2, 10)
-
-    """
-
-    @property
-    def parameter_names(self) -> list[str]:
-        """Names of parameters used by the model: `center_index`, `sigma`."""
-        return ["center_index", "sigma"]
-
-    @doc
-    def __call__(self, stimulus: CFStimulus, parameters: pd.DataFrame, dtype: str | None = None) -> Tensor:
-        """
-        Predict the model response for a stimulus with a distance matrix.
-
-        Parameters
-        ----------
-        %(stimulus_cf)s
-        %(parameters)s
-        %(dtype)s
-
-        Returns
-        -------
-        Tensor
-            Model predictions of shape `(num_units, num_rows)` and dtype `dtype`.
-            `num_units` is the number of rows in `parameters` and `num_rows` is the number of rows in the stimulus
-            distance matrix.
-
-        """
-        dtype = get_dtype(dtype)
-        # Distance matrix is numpy array so we also create a numpy array to safely index
-        # The dtype is only used for indexing so it can be hardcoded
-        center_index = np.asarray(parameters[["center_index"]], dtype=np.int32)[:, 0]
-        sigma = convert_parameters_to_tensor(parameters[["sigma"]], dtype=dtype)
-        distance_matrix = ops.convert_to_tensor(stimulus.distance_matrix[center_index], dtype=dtype)
-
-        sigma_squared = ops.square(sigma)
-
-        # Gaussian response
-        resp = ops.square(distance_matrix)
-        resp /= 2.0 * sigma_squared
-
-        # Divide by volume to normalize (only two dimensions, so exponent cancels out)
-        volume = ops.sqrt(2.0 * math.pi * sigma_squared)
-
-        return ops.exp(-resp) / volume
 
 
 class Gaussian2DPRFModel(SimplePRFModel):
