+++ tests/fitters/test_grid.py
@@ -7,7 +7,7 @@ import pytest
 from prfmodel.fitters.grid import GridFitter
 from prfmodel.fitters.grid import GridHistory
 from prfmodel.fitters.grid import InfiniteLossWarning
-from prfmodel.models.gaussian import Gaussian2DPRFModel
+from prfmodel.models.prf.gaussian import Gaussian2DPRFModel
 from prfmodel.stimuli import PRFStimulus
 from tests.conftest import TestSetup
 from tests.conftest import parametrize_impulse_model
