+++ tests/fitters/test_least_squares.py
@@ -6,8 +6,8 @@ import pytest
 from pytest_regressions.dataframe_regression import DataFrameRegressionFixture
 from prfmodel.fitters.least_squares import LeastSquaresFitter
 from prfmodel.fitters.least_squares import LeastSquaresHistory
-from prfmodel.models.dog import DoG2DPRFModel
-from prfmodel.models.gaussian import Gaussian2DPRFModel
+from prfmodel.models.prf.dog import DoG2DPRFModel
+from prfmodel.models.prf.gaussian import Gaussian2DPRFModel
 from prfmodel.stimuli import PRFStimulus
 from tests.conftest import PRFStimulusSetup
 from tests.conftest import TestSetup
