+++ tests/models/prf/test_divisive_normalization.py
@@ -6,10 +6,10 @@ import pytest
 from pytest_regressions.num_regression import NumericRegressionFixture
 from prfmodel.impulse import DerivativeTwoGammaImpulse
 from prfmodel.impulse.base import BaseImpulse
-from prfmodel.models.div_norm import DivNormGaussian2DPRFModel
-from prfmodel.models.div_norm import DivNormPRFModel
-from prfmodel.models.div_norm import init_dn_from_gaussian
-from prfmodel.models.gaussian import Gaussian2DPRFResponse
+from prfmodel.models.prf.div_norm import DivNormGaussian2DPRFModel
+from prfmodel.models.prf.div_norm import DivNormPRFModel
+from prfmodel.models.prf.div_norm import init_dn_from_gaussian
+from prfmodel.models.prf.gaussian import Gaussian2DPRFResponse
 from prfmodel.scaling import DivNormAmplitude
 from prfmodel.scaling.base import BaseScaling
 from prfmodel.stimuli.prf import PRFStimulus
